UI_Code/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pyserial==3.5
python-socketio==5.8.0
eventlet==0.33.3
pyecharts==2.0.4